fix: Handle `podman` datestamps in `Makefile`
This change ensures that we get an identical datestamp format out of
both `docker` and `podman` when we check for container image age in the
`Makefile`.

`podman` uses the standard Go `time.Time` type which renders differently
to whatever custom interace `docker` uses. It'd be lovely to simply use
a Go template like `{{.Created.Unix}}` but unfortunately the `docker`
datestamp type doesn't have such a method so instead we have to abuse
the presence of the `json` encoder to get a standard output format which
POSIX `date(1)` should understand. We also have an unfortunate pipeline
step where we have to strip out the quotes around the extracted JSON
string since `date` barfs if we leave them there.