Update apt to newer version to avoid bugs
Debian Snapshot seems to be struggling with older versions of apt.
This commit uses some hacks to get a known-working version of apt,
if necessary, and installs it before most other operations.

This seems to fix a lot of the 'Undetermined errors' that apt would
throw when trying to pull Debian Snapshot debs.

Signed-off-by: Luke Mondy <luke.mondy@data61.csiro.au>