Migrate logic away from user.Dockerfile
And into a bash script. This is much cleaner for reading and linting,
and as such, much easier to add logic to. In this case, more complex
logic on group creation has been added. If group creation fails, we now
abandon trying to setup the group, which is mostly a nice-to-have.

Signed-off-by: Luke Mondy <luke.mondy@data61.csiro.au>