Use Debian-provided riscv toolchain
This uses less space in the seL4 base image, which means we can remove
the separate -riscv images. We keep dummy -riscv images for now, so
that CI that expects these to exist continues to work.

This gives us the opportunity to replace most of the -riscv logic in
the Makefile with -tex logic instead (but -riscv is still supported in
the ./build.sh script).

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>