Switch docker builds to Debian bullseye
bullseye has been released as the next "stable" and the old docker
build does not work any more.

This commit switches the entire build over to bullseye only, which
also happens to simplify the build scripts and upgrade a number of
tool chain packages. It also means the `riscv` variant is removed,
and included in the main images. In particular, we now have in the
sel4 image:

- gcc-10
- clang-11
- riscv gcc-8 + qemu 5.2.9

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>