github: separate out the l4v build on PR
As in the main deployment build, the l4v image overflows the runner's
disk space if done in the same job as the rest of the images. This
means we're now building sel4/camkes twice, but not cakeml/cogent/rust,
which hopefully frees up enough space.

This refactor also removes the snapshot date setting, which is not
necessary for PR builds. Removing it makes the build slightly more
reliable, because the Debian snapshot repositories sometimes seem to
have network problems. We still have reproducible builds in the main
deployment action.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>