remove Debian snapshots as default
- make not using snapshots the default
- remove snapshots from CI build

While the snapshots in theory make the build more reproducible, the
snapshot repos is notoriously unreliable in CI making the build fail
more often than it succeeds.

Since we publish and tag the CI builds, if somebody wants to use the
exact same base containers, they can use the tagged image -- reproducing
the image build from scratch is not necessary for that scenario.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>